refactor custom-element export
