+++ CHANGELOG.md
@@ -1,0 +1,4 @@
+## v4.0:
+
+- [breaking] changed signature of `registerSvelecte`
+- [breaking] `registerSvelecte` is now available from `svelecte` itself, not from `svelecte/component`
