+++ index.js
@@ -1,5 +1,5 @@
 import Svelecte, { addFormatter, config } from './src/Svelecte.svelte';
 
 export default Svelecte;
-
 export { addFormatter, config }
+export { registerSvelecte } from './src/lib/custom-element';
